+++ src/main/java/com/example/carrentsys/repository/ClientRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClientRepository extends CrudRepository<Client, Long> {
+public interface ClientRepository extends CrudRepository<Client, Integer> {
     Client findByUsername(String username);
 
     boolean existsByUsername(String username);
